kukui: move damu (asus cm3 flip) to tested
recently i had access to an asus cm3 flip (damu) for some time and was
able to test it with the latest image and to at least partially fix the
audio setup for it (only speaker and mic are working so far). it was
the 8gb ram version of it btw.