7c-woa: get back the dispay after drm init without scaling
currently the display shows wild noise from when drm takes over (i.e.
gpu etc. gets initialized) and some xrandr scaling was required to bring
it back to normal - this changes it to an xrandr off and auto which
brings it back to life as well without having to set any kind of scaling